+++ src/PhpBrew/ExtensionInterface.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace PhpBrew;
+
+interface ExtensionInterface
+{
+    public function __construct($name, $logger);
+    public function install($version, array $options = array());
+    public function enable();
+    public function disable();
+    public function isLoaded();
+    public function isAvailable();
+}