+++ src/PhpBrew/Utils.php
@@ -22,44 +22,6 @@ class Utils
         }
     }
 
-    static function get_extension_config_path($extname)
-    {
-        // create extension config file
-        $path = Config::getCurrentPhpConfigScanPath() . DIRECTORY_SEPARATOR . $extname . '.ini';
-        if ( ! file_exists( dirname($path) ) ) {
-            mkdir(dirname($path),0755,true);
-        }
-        return $path;
-    }
-
-    static function enable_extension($extname, $zendpath = '')
-    {
-        $extname = strtolower($extname);
-        // create extension config file
-        $configPath = self::get_extension_config_path($extname);
-
-        if ( file_exists($configPath) ) {
-            $lines = file($configPath);
-            foreach( $lines as &$line ) {
-                if ( preg_match('#^;\s*((?:zend_)?extension\s*=.*)#', $line, $regs ) ) {
-                    $line = $regs[1];
-                }
-            }
-            file_put_contents($configPath, join("\n", $lines) );
-            return $configPath;
-        } else {
-            if( $zendpath ) {
-                $content = "zend_extension=$zendpath";
-            } else {
-                $content = "extension=$extname.so";
-            }
-            file_put_contents($configPath,$content);
-            return $configPath;
-        }
-        return false;
-    }
-
-
     /**
      * Find bin from prefix list
      */
@@ -196,6 +158,12 @@ class Utils
         }
     }
 
+    /**
+     * Find executable binary by PATH environment.
+     *
+     * @param string $bin binary name
+     * @return string the path
+     */
     static function findbin($bin)
     {
         $path = getenv('PATH');
@@ -203,6 +171,9 @@ class Utils
         foreach( $paths as $path ) {
             $f = $path . DIRECTORY_SEPARATOR . $bin;
             if( file_exists($f) ) {
+                while ( is_link($f) ) {
+                    $f = readlink($f);
+                }
                 return $f;
             }
         }