+++ tests/PhpBrew/UtilsTest.php
@@ -17,5 +17,10 @@ class UtilsTest extends PHPUnit_Framework_TestCase
         ok( Utils::find_lib_prefix('icu/pkgdata.inc','icu/Makefile.inc') );
         ok( Utils::find_include_prefix('openssl/opensslv.h') );
     }
+
+    public function testFindbin() {
+        ok(Utils::findbin('ls'));
+        ok(Utils::findbin('psql'));
+    }
 }